moving parallel gridcalc
